Code Cleanup
- Various code style cleanup and improvements, using latest C# features.

+++ frmGridSize.cs
@@ -1,10 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
 using System.Windows.Forms;
 using SpriteEditor.Properties;
 
@@ -25,16 +19,16 @@ namespace SpriteEditor
 
         private void frmGridSize_Load(object sender, EventArgs e)
         {
-            nudGridWidth.Value = (decimal) Settings.Default.GridWidth;
-            nudGridHeight.Value = (decimal) Settings.Default.GridHeight;
+            nudGridWidth.Value = Settings.Default.GridWidth;
+            nudGridHeight.Value = Settings.Default.GridHeight;
             cdGridColor.Color = Settings.Default.GridColor;
             btnGridColor.BackColor = cdGridColor.Color;
         }
 
         private void btnOK_Click(object sender, EventArgs e)
         {
-            Settings.Default.GridWidth = (int) nudGridWidth.Value;
-            Settings.Default.GridHeight = (int) nudGridHeight.Value;
+            Settings.Default.GridWidth = (int)nudGridWidth.Value;
+            Settings.Default.GridHeight = (int)nudGridHeight.Value;
             Settings.Default.GridColor = cdGridColor.Color;
             Hide();
         }
